Implement gpu-raytracer
The rendering is now done on the gpu, using wgpu-rs. winit is used to
open the window and handle events. Additionally, a simple gui is
rendered with imgui-rs. For simplicity, the renderer supports only
static models and the Entity-Component-Code was (temporarily)
removed. The renderer only supports a single directional light.